feat: modify benchmark
