+++ template/index.sass
@@ -91,6 +91,7 @@ a
 
     .nav-right
         flex: 1
+        position: relative
         box-shadow: rgb(125 121 121) -2px 0px 8px
 
         iframe
